+++ ChozoEngine/src/Chozo/Scene/Entity.h
@@ -49,7 +49,7 @@ namespace Chozo
 
         operator bool() const { return m_EntityHandle != entt::null; }
         operator entt::entity() const { return m_EntityHandle; }
-        operator uint32_t() const { return (uint32_t)m_EntityHandle; }
+        operator uint64_t() const { return (uint64_t)m_EntityHandle; }
         bool operator==(const Entity& other) const {
             return m_EntityHandle == other.m_EntityHandle && m_Scene == other.m_Scene;
         }
